Moving the documentation to Hugo (#1177)
* Moving to Hugo based documentation

Adding semi static documentation based on Hugo

Signed-off-by: Gergely Csatari <gergely.csatari@nokia.com>

* Adding versions and examples to the function pages

Every function page has now a drop down for the versions of that function and a
drop down for the examples.

Note: The same examples are shown for all the versions even if the given version of the
function does not fully supports the example.

Examples are copied from the examples folder of the repo.

Signed-off-by: Gergely Csatari <gergely.csatari@nokia.com>

---------

Signed-off-by: Gergely Csatari <gergely.csatari@nokia.com>